♻️ Refactor student management components: - Removed SearchBar and AdaptiveTable components, replacing them with SearchFilterBar and DataTable. - Updated TutorStudentsTemplate to integrate new components and handle student addition through a new EntityForm. - Enhanced modal functionality for adding students and viewing details.
